new code architecture (close #11)
