+++ src/apple/gpu.cpp
@@ -1,0 +1,42 @@
+// Copyright Leon Freist
+// Author Leon Freist <freist@informatik.uni-freiburg.de>
+
+#include "hwinfo/platform.h"
+
+#ifdef HWINFO_APPLE
+
+#include <vector>
+#include <string>
+#include <regex>
+
+#include "hwinfo/gpu.h"
+
+namespace hwinfo {
+
+// _____________________________________________________________________________________________________________________
+std::string GPU::getVendor() {
+  // TODO: implement
+  return "<unknown>";
+}
+
+// _____________________________________________________________________________________________________________________
+std::string GPU::getName() {
+  // TODO: implement
+  return "<unknown>";
+}
+
+// _____________________________________________________________________________________________________________________
+std::string GPU::getDriverVersion() {
+  // TODO: implement
+  return "<unknown>";
+}
+
+// _____________________________________________________________________________________________________________________
+int64_t GPU::getMemory_Bytes() {
+  // TODO: implement
+  return "<unknown>";
+}
+
+}  // namespace hwinfo
+
+#endif  // HWINFO_APPLE
