+++ src/apple/os.cpp
@@ -1,0 +1,63 @@
+// Copyright (c) Leon Freist <freist@informatik.uni-freiburg.de>
+// This software is part of HWBenchmark
+
+#include "hwinfo/platform.h"
+
+#ifdef HWINFO_APPLE
+
+#include <sstream>
+#include <string>
+
+#include <sys/sysctl.h>
+
+#include "hwinfo/os.h"
+
+namespace hwinfo {
+
+// _____________________________________________________________________________________________________________________
+std::string OS::getFullName() {
+  string os_name;
+  os_name.resize(1024);
+  size_t size = sizeof(os_name);
+  if (sysctlbyname("kern.os", os_name.data(), &size, nullptr, 0) == 0) {
+    return os_name;
+  }
+  return "macOS <unknown version>";
+}
+
+// _____________________________________________________________________________________________________________________
+std::string OS::getName() {
+  string os_name;
+  os_name.resize(1024);
+  size_t size = sizeof(os_name);
+  if (sysctlbyname("kern.os", os_name.data(), &size, nullptr, 0) == 0) {
+    return os_name;
+  }
+  return "macOS";
+}
+
+// _____________________________________________________________________________________________________________________
+std::string OS::getVersion() {
+  std::string os_name;
+  os_name.resize(1024);
+  size_t size = sizeof(os_name);
+  if (sysctlbyname("kern.osrelease", os_name.data(), &size, nullptr, 0) == 0) {
+    return os_name;
+  }
+  return "<unknown version>";
+}
+
+// _____________________________________________________________________________________________________________________
+std::string OS::getKernel() {
+  // TODO: implement
+  return "<unknown>";
+}
+
+// _____________________________________________________________________________________________________________________
+bool OS::getIs64bit() {
+  return true;
+}
+
+}  // namespace hwinfo
+
+#endif  // HWINFO_APPLE
