+++ src/ram.cpp
@@ -69,86 +69,4 @@ int64_t RAM::totalSize_Bytes() {
   return _totalSize_Bytes;
 }
 
-// _____________________________________________________________________________________________________________________
-std::string RAM::getManufacturer() {
-#if defined(_WIN32) || defined(_WIN64)
-  std::vector<const wchar_t*> names{};
-  wmi::queryWMI("WIN32_PhysicalMemory", "Manufacturer", names);
-  auto ret = names[0];
-  std::wstring tmp(ret);
-  return {tmp.begin(), tmp.end()};
-#else
-  return "<unknown>";
-#endif
-}
-
-// _____________________________________________________________________________________________________________________
-std::string RAM::getName() {
-#if defined(_WIN32) || defined(_WIN64)
-  std::vector<const wchar_t*> names{};
-  wmi::queryWMI("WIN32_PhysicalMemory", "Name", names);
-  auto ret = names[0];
-  std::wstring tmp(ret);
-  return {tmp.begin(), tmp.end()};
-#else
-  return "<unknown>";
-#endif
-}
-
-// _____________________________________________________________________________________________________________________
-std::string RAM::getModel() {
-#if defined(_WIN32) || defined(_WIN64)
-  std::vector<const wchar_t*> names{};
-  wmi::queryWMI("WIN32_PhysicalMemory", "PartNumber", names);
-  auto ret = names[0];
-  std::wstring tmp(ret);
-  return {tmp.begin(), tmp.end()};
-#else
-  return "<unknown>";
-#endif
-}
-
-// _____________________________________________________________________________________________________________________
-std::string RAM::getSerialNumber() {
-#if defined(_WIN32) || defined(_WIN64)
-  std::vector<const wchar_t*> names{};
-  wmi::queryWMI("WIN32_PhysicalMemory", "SerialNumber", names);
-  auto ret = names[0];
-  std::wstring tmp(ret);
-  return {tmp.begin(), tmp.end()};
-#else
-  return "<unknown>";
-#endif
-}
-
-// _____________________________________________________________________________________________________________________
-int64_t RAM::getTotalSize_Bytes() {
-#if defined(unix) || defined(__unix) || defined(__unix__)
-  long pages = sysconf(_SC_PHYS_PAGES);
-  long page_size = sysconf(_SC_PAGESIZE);
-  if ((pages > 0) && (page_size > 0)) {
-    return pages * page_size;
-  }
-  return -1;
-#elif defined(__APPLE__)
-  int64_t memsize = 0;
-  size_t size = sizeof(memsize);
-  if (sysctlbyname("hw.memsize", &memsize, &size, nullptr, 0) == 0) {
-    return memsize;
-  }
-  return -1;
-#elif defined(_WIN32) || defined(_WIN64)
-  MEMORYSTATUSEX status;
-  status.dwLength = sizeof(status);
-  GlobalMemoryStatusEx(&status);
-  return static_cast<int64_t>(status.ullTotalPhys);
-#else
-  return -1;
-#endif
-}
-
-// _____________________________________________________________________________________________________________________
-
-
-// _____________________________________________________________________________________________________________________
 }  // namespace hwinfo
